Create delete comment component
